V1.0

+++ species_recgonize.py
@@ -19,6 +19,39 @@ params = urllib.parse.urlencode({
     'predictMode': 'classifyAndDetect',
 })
 
+def rendering_box(l, img):
+    image = mpimg.imread(img)
+    #count = 0
+    dpi = 100
+
+    print(image.shape)
+    imageHeight, imageWidth = image.shape[0:2]
+    figsize = imageWidth / float(dpi), imageHeight / float(dpi)
+    fig = plt.figure(figsize=figsize)
+    ax = plt.axes([0,0,1,1])
+
+        # Display the image
+    ax.imshow(image)
+    ax.set_axis_off()
+    
+    
+    rect = patches.Rectangle((float(l['x_min']),float(l['y_min'])),float(l['x_max']) - float(l['x_min']),float(item['y_max']) - float(item['y_min']),linewidth=3,edgecolor=color_list[count],facecolor='none')
+        # Add the patch to the Axes
+    ax.add_patch(rect)
+
+    # This is magic goop that removes whitespace around image plots (sort of)        
+    plt.subplots_adjust(top = 1, bottom = 0, right = 1, left = 0, hspace = 0, wspace = 0)
+    plt.margins(0,0)
+    ax.xaxis.set_major_locator(ticker.NullLocator())
+    ax.yaxis.set_major_locator(ticker.NullLocator())
+    ax.axis('tight')
+    ax.set(xlim=[0,imageWidth],ylim=[imageHeight,0],aspect=1)
+    plt.axis('off')                
+    
+    # plt.savefig(outputFileName, bbox_inches='tight', pad_inches=0.0, dpi=dpi, transparent=True)
+    plt.savefig(img, dpi=dpi, transparent=True)
+    print ('done!')
+
 def species_recgonize():
     while True:
         cursor = db.cursor()
@@ -67,38 +100,40 @@ def species_recgonize():
                     print('Error: unable to fetch the joblist')
                 
                 if species in s:
-                    sql = "SELECT USERID FROM DEVICEINFO WHERE DEVICEID='%s' " %deviceId
-
-                    try:
-                        cursor.execute(sql)
-                        result = cursor.fetchone()
-                        userId = result[0]
-                    
-                    except:
-                        print('Error: unable to fetch userid')
-
-                    sql = "INSERT INTO IMGINFO (IMGNAME, USERID, DEVICEID, TIMESTAMP, JOB, CONFIDENCE) \
-                        VALUES ('%s', '%s', '%s', '%s', '%s', '%s')" %(row[0], userId, deviceId, timestamp, species, confidence)
-                    
-                    cursor.execute(sql)
-                    db.commit()
-                    try:
-                        
-                        status = 1
-                        
-                    except:
-                        print('Error: unable to insert data')
-                        status = 0
-                    
-                    if status == 1:
-                        sql = "DELETE FROM IMGRECEIVED WHERE IMGNANE = '%s' " %row[0]
+                    if d['bboxes'] != []:
+                        rendering_box(d['bboxes'], 'http://40.112.164.41:5000/' + str(n))
+                        sql = "SELECT USERID FROM DEVICEINFO WHERE DEVICEID='%s' " %deviceId
 
                         try:
                             cursor.execute(sql)
-                            db.commit()
-
+                            result = cursor.fetchone()
+                            userId = result[0]
+                        
                         except:
-                            print('delete from imgreceived failed')
+                            print('Error: unable to fetch userid')
+
+                        sql = "INSERT INTO IMGINFO (IMGNAME, USERID, DEVICEID, TIMESTAMP, JOB, CONFIDENCE) \
+                            VALUES ('%s', '%s', '%s', '%s', '%s', '%s')" %(row[0], userId, deviceId, timestamp, species, confidence)
+                        
+                        cursor.execute(sql)
+                        db.commit()
+                        try:
+                            
+                            status = 1
+                            
+                        except:
+                            print('Error: unable to insert data')
+                            status = 0
+                        
+                        if status == 1:
+                            sql = "DELETE FROM IMGRECEIVED WHERE IMGNANE = '%s' " %row[0]
+
+                            try:
+                                cursor.execute(sql)
+                                db.commit()
+
+                            except:
+                                print('delete from imgreceived failed')
                 
                 else:
                     os.remove('./' + n)
